backend/requirements.txt
# LogGuard AI — FastAPI backend (Python 3.11–3.13 recommended; 3.14+ when wheels exist)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
huggingface_hub>=0.27.0
httpx>=0.27.0
pydantic>=2.11.0
pydantic-settings>=2.6.0
scikit-learn>=1.5.0
numpy>=1.26.0
reportlab>=4.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
